requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
distlib==0.3.5
Django==4.0.6
django-jazzmin==2.5.0
django-ninja==0.19.0
django-rest-framework==0.1.0
django-shortcuts==1.6
djangorestframework==3.13.1
dnspython==2.2.1
docopt==0.6.2
ecdsa==0.18.0
email-validator==1.2.1
filelock==3.7.1
idna==3.3
Pillow==9.2.0
pipenv==2022.7.4
pipreqs==0.4.11
platformdirs==2.5.2
pyasn1==0.4.8
pycodestyle==2.8.0
pydantic==1.9.1
python-jose==3.3.0
pytz==2022.1
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
virtualenv==20.15.1
virtualenv-clone==0.5.7
yarg==0.1.9
